Add simple eyeball-able test to check that it works as advertised.
I don't see the value in trying to figure out how to capture all
the output from both the shell and node to compare the bytes. If
I happen to learn that later I'll update the test to report a
pass/fail automatically.

In the meantime, it is pretty easy to look at and see that the
output looks the same.